+++ packages/perseus-core/src/parse-perseus-json/perseus-parsers/plotter-widget.ts
@@ -1,6 +1,3 @@
-// TODO(LEMS-4224): don't import from outside of the parser
-// eslint-disable-next-line import/no-restricted-paths
-import {plotterPlotTypes} from "../../data-schema";
 import {
     constant,
     object,
@@ -15,12 +12,20 @@ import {defaulted} from "../general-purpose-parsers/defaulted";
 
 import {parseWidget} from "./widget";
 
+const plotterPlotTypes = enumeration(
+    "bar",
+    "line",
+    "pic",
+    "histogram",
+    "dotplot",
+);
+
 export const parsePlotterWidget = parseWidget(
     constant("plotter"),
     object({
         labels: array(string),
         categories: array(string),
-        type: enumeration(...plotterPlotTypes),
+        type: plotterPlotTypes,
         maxY: number,
         // The default value for scaleY comes from plotter.tsx.
         // See parse-perseus-json/README.md for why we want to duplicate the