+++ packages/perseus/src/components/sortable.tsx
@@ -2,6 +2,7 @@
 /* eslint-disable @typescript-eslint/no-invalid-this, react/no-unsafe */
 import * as PerseusLinter from "@khanacademy/perseus-linter";
 import {CircularSpinner} from "@khanacademy/wonder-blocks-progress-spinner";
+import {border, semanticColor} from "@khanacademy/wonder-blocks-tokens";
 import {StyleSheet, css} from "aphrodite";
 import $ from "jquery";
 import * as React from "react";
@@ -917,8 +918,8 @@ const styles = StyleSheet.create({
 
     card: {
         boxSizing: "border-box",
-        background: "#fff",
-        border: "1px solid #ddd",
+        background: semanticColor.core.background.base.default,
+        border: `${border.width.thin} solid ${semanticColor.core.border.neutral.subtle}`,
         borderRadius: 4,
         cursor: "pointer",
         minWidth: 25,
@@ -932,8 +933,8 @@ const styles = StyleSheet.create({
     },
 
     placeholder: {
-        background: "#ddd",
-        border: "1px solid #ccc",
+        background: semanticColor.core.background.neutral.subtle,
+        border: `${border.width.thin} solid ${semanticColor.core.border.neutral.subtle}`,
     },
 
     draggable: {
@@ -955,13 +956,14 @@ const styles = StyleSheet.create({
     },
 
     dragging: {
-        background: "#ffedcd",
+        background: semanticColor.core.background.instructive.subtle,
+        border: `${border.width.thin} solid ${semanticColor.core.border.instructive.default}`,
         opacity: 0.8,
     },
 
     disabled: {
         backgroundColor: "inherit",
-        border: "1px solid transparent",
+        border: `${border.width.thin} solid ${semanticColor.core.border.disabled.subtle}`,
         cursor: "default",
     },
 });