+++ packages/perseus/src/widgets/image/__docs__/image-initial-state-regression.stories.tsx
@@ -343,9 +343,6 @@ export const ImageWithoutWidthOrHeightLarge: Story = {
 
 export const TallAnimatedGif: Story = {
     decorators: [imageRendererDecorator],
-    globals: {
-        featureFlags: ["image-widget-upgrade-gif-controls"],
-    },
     args: {
         backgroundImage: animatedGifPortrait,
         alt: animatedGifPortraitAlt,
@@ -414,9 +411,6 @@ export const PngImage: Story = {
 
 export const MobileAnimatedGif: Story = {
     decorators: [imageRendererDecorator, mobileDecorator],
-    globals: {
-        featureFlags: ["image-widget-upgrade-gif-controls"],
-    },
     args: {
         backgroundImage: animatedGifLandscape,
         alt: animatedGifLandscapeAlt,
@@ -426,9 +420,6 @@ export const MobileAnimatedGif: Story = {
 
 export const NonAnimatedGif: Story = {
     decorators: [imageRendererDecorator],
-    globals: {
-        featureFlags: ["image-widget-upgrade-gif-controls"],
-    },
     args: {
         backgroundImage: nonAnimatedGif,
         alt: nonAnimatedGifAlt,
@@ -466,7 +457,7 @@ export const MarkdownTableWithImageWidgets: Story = {
                             }),
                             "image 2": generateImageWidget({
                                 options: generateImageOptions({
-                                    backgroundImage: earthMoonImage,
+                                    backgroundImage: scienceImage,
                                     alt: "Earth and Moon",
                                 }),
                             }),
@@ -494,7 +485,7 @@ export const MarkdownTableWithMarkdownImages: Story = {
             <div style={{width: 600}}>
                 <QuestionRendererForStories
                     question={generateTestPerseusRenderer({
-                        content: `| col 1 | col 2 | col 3 |\n| --- | --- | --- |\n| ![Fresco painting](${frescoImage.url}) | ![Earth and Moon](${earthMoonImage.url}) | ![Graphie image](${graphieImage.url}) |`,
+                        content: `| col 1 | col 2 | col 3 |\n| --- | --- | --- |\n| ![Fresco painting](${frescoImage.url}) | ![${scienceImageAlt}](${scienceImage.url}) | ![Graphie image](${graphieImage.url}) |`,
                         widgets: {},
                     })}
                 />