+++ packages/perseus-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/perseus-core
 
+## 27.4.0
+
+### Minor Changes
+
+-   [#3725](https://github.com/Khan/perseus/pull/3725) [`9c601da23f`](https://github.com/Khan/perseus/commit/9c601da23fb67a0de9af0694fe0f877998b70c88) Thanks [@EmiliaPalaghita](https://github.com/EmiliaPalaghita)! - Add new JSON field "showPointLabels" for interactive graphs. When set, every movable point gets a visible on-canvas label driven by the matching `pointLabels[i]` entry. `pointLabels` is required whenever `showPointLabels` is true (enforced by the interactive-graph-widget-error lint rule).
+
+### Patch Changes
+
+-   [#3742](https://github.com/Khan/perseus/pull/3742) [`d15912407c`](https://github.com/Khan/perseus/commit/d15912407c8bac67bbd68a13487d7ad453da2f81) Thanks [@nishasy](https://github.com/nishasy)! - [Image] Remove gif flag, make gif controls permanent
+
+*   [#3726](https://github.com/Khan/perseus/pull/3726) [`ec0ce8d61a`](https://github.com/Khan/perseus/commit/ec0ce8d61a8f8565acba672fe249723d025f88fc) Thanks [@handeyeco](https://github.com/handeyeco)! - Copy data that was once shared into perseus-parser
+
 ## 27.3.0
 
 ### Minor Changes